Tests: test_scipy_special() added.
This test exercises a new hidden import solving #1908. For maintainability, all
SciPy tests have also been moved from "test_libraries.py" into a new
"test_hooks/test_scipy.py" script. Thanks to scipy/scipy#5461, these tests have
also been marked as failing under Windows.

Sadly, this test has also exposed #1919, a new non-trivial "ModuleGraph" issue.
Until addressed, the test_scipy_special() test must be unconditionally skipped.
