+++ tests/functional/test_libraries.py
@@ -491,35 +491,6 @@ def test_scapy3(pyi_builder):
         """)
 
 
-@xfail(is_darwin, reason='Issue #1895.')
-@importorskip('scipy')
-def test_scipy(pyi_builder):
-    pyi_builder.test_source(
-        """
-        from distutils.version import LooseVersion
-
-        # Test top-level SciPy importability.
-        import scipy
-        from scipy import *
-
-        # Test hooked SciPy modules.
-        import scipy.io.matlab
-        import scipy.sparse.csgraph
-
-        # Test problematic SciPy modules.
-        import scipy.linalg
-        import scipy.signal
-
-        # SciPy >= 0.16 privatized the previously public "scipy.lib" package as
-        # "scipy._lib". Since this package is problematic, test its
-        # importability regardless of SciPy version.
-        if LooseVersion(scipy.__version__) >= LooseVersion('0.16.0'):
-            import scipy._lib
-        else:
-            import scipy.lib
-        """)
-
-
 @importorskip('sqlalchemy')
 def test_sqlalchemy(pyi_builder):
     pyi_builder.test_source(
